feat: Implement getYearLordDignityDescription and fix type mismatch
Implemented the `getYearLordDignityDescription` function in `VarshaphalaCalculator.kt` to resolve the unresolved reference error. This function provides a detailed description of the Year Lord's dignity based on its position in the solar return chart.

Fixed the `Type mismatch` error in `VarshaphalaScreen.kt` by explicitly converting the `Double` value to a `Float` for the `LinearProgressIndicator`.

+++ app/src/main/java/com/astro/storm/ui/screen/VarshaphalaScreen.kt
@@ -1175,7 +1175,7 @@ private fun PlanetBalaRow(bala: PanchaVargiyaBala) {
         )
 
         LinearProgressIndicator(
-            progress = { (bala.total / 20f).coerceIn(0f, 1f) },
+            progress = { (bala.total.toFloat() / 20f).coerceIn(0f, 1f) },
             modifier = Modifier
                 .weight(1f)
                 .height(8.dp)
